ENG-18786: Eliminate the bulk of ExportOnServerVerifier
The only code in use in ExportOnServerVerifier is the static
    VerifyRows()
method.  Moved this to new RowVerifier class.
